+++ src/main.rs
@@ -13,8 +13,8 @@ use clap::{Parser, Subcommand};
 #[cfg(feature = "scaffolding")]
 use maturin::{ci::GenerateCI, init_project, new_project, GenerateProjectOptions};
 use maturin::{
-    develop, write_dist_info, Abi3Version, BridgeModel, BuildOptions, CargoOptions, DevelopOptions,
-    PathWriter, PlatformTag, PythonInterpreter, Target, TargetTriple,
+    develop, write_dist_info, BridgeModel, BuildOptions, CargoOptions, DevelopOptions, PathWriter,
+    PythonInterpreter, Target, TargetTriple,
 };
 #[cfg(feature = "schemars")]
 use maturin::{generate_json_schema, GenerateJsonSchemaOptions};
@@ -280,35 +280,13 @@ fn pep517(subcommand: Pep517Command) -> Result<()> {
                 .editable(false)
                 .build()?;
 
-            // Since afaik all other PEP 517 backends also return linux tagged wheels, we do so too
-            let tags = match context.bridge() {
-                BridgeModel::PyO3(bindings) | BridgeModel::Bin(Some(bindings)) => {
-                    match bindings.abi3 {
-                        Some(Abi3Version::Version(major, minor)) => {
-                            let platform = context.get_platform_tag(&[PlatformTag::Linux])?;
-                            vec![format!("cp{major}{minor}-abi3-{platform}")]
-                        }
-                        Some(Abi3Version::CurrentPython) => {
-                            let interp = &context.interpreter[0];
-                            let platform = context.get_platform_tag(&[PlatformTag::Linux])?;
-                            vec![format!(
-                                "cp{major}{minor}-abi3-{platform}",
-                                major = interp.major,
-                                minor = interp.minor
-                            )]
-                        }
-                        None => {
-                            vec![context.interpreter[0].get_tag(&context, &[PlatformTag::Linux])?]
-                        }
-                    }
-                }
-                BridgeModel::Bin(None) | BridgeModel::Cffi | BridgeModel::UniFfi => {
-                    context.get_universal_tags(&[PlatformTag::Linux])?.1
-                }
-            };
-
             let mut writer = PathWriter::from_path(metadata_directory);
-            write_dist_info(&mut writer, &context.metadata24, &tags)?;
+            write_dist_info(
+                &mut writer,
+                &context.project_layout.project_root,
+                &context.metadata24,
+                &context.tags_from_bridge()?,
+            )?;
             println!("{}", context.metadata24.get_dist_info_dir().display());
         }
         Pep517Command::BuildWheel {